Move LocalizedContentItemIndexProvider to OC.ContentLocalization.Module
