+++ src/OrchardCore.Modules/OrchardCore.ContentLocalization/Indexes/LocalizedContentItemIndexProvider.cs
@@ -1,0 +1,96 @@
+using Microsoft.Extensions.DependencyInjection;
+using OrchardCore.ContentLocalization.Models;
+using OrchardCore.ContentLocalization.Records;
+using OrchardCore.ContentManagement.Handlers;
+using OrchardCore.ContentManagement.Metadata;
+using OrchardCore.ContentManagement;
+using OrchardCore.Data;
+using YesSql.Indexes;
+
+namespace OrchardCore.ContentLocalization.Indexes;
+
+public class LocalizedContentItemIndexProvider : ContentHandlerBase, IIndexProvider, IScopedIndexProvider
+{
+    private readonly IServiceProvider _serviceProvider;
+    private readonly HashSet<ContentItem> _itemRemoved = [];
+    private readonly HashSet<string> _partRemoved = [];
+    private IContentDefinitionManager _contentDefinitionManager;
+
+    public LocalizedContentItemIndexProvider(IServiceProvider serviceProvider)
+    {
+        _serviceProvider = serviceProvider;
+    }
+
+    public override Task RemovedAsync(RemoveContentContext context)
+    {
+        if (context.NoActiveVersionLeft)
+        {
+            var part = context.ContentItem.As<LocalizationPart>();
+
+            if (part != null)
+            {
+                _itemRemoved.Add(context.ContentItem);
+            }
+        }
+
+        return Task.CompletedTask;
+    }
+
+    public override async Task PublishedAsync(PublishContentContext context)
+    {
+        var part = context.ContentItem.As<LocalizationPart>();
+
+        // Validate that the content definition contains this part, this prevents indexing parts
+        // that have been removed from the type definition, but are still present in the elements.            
+        if (part != null)
+        {
+            // Lazy initialization because of ISession cyclic dependency.
+            _contentDefinitionManager ??= _serviceProvider.GetRequiredService<IContentDefinitionManager>();
+
+            // Search for this part.
+            var contentTypeDefinition = await _contentDefinitionManager.GetTypeDefinitionAsync(context.ContentItem.ContentType);
+            if (!contentTypeDefinition.Parts.Any(ctd => ctd.Name == nameof(LocalizationPart)))
+            {
+                context.ContentItem.Remove<LocalizationPart>();
+                _partRemoved.Add(context.ContentItem.ContentItemId);
+            }
+        }
+    }
+
+    public string CollectionName { get; set; }
+    public Type ForType() => typeof(ContentItem);
+    public void Describe(IDescriptor context) => Describe((DescribeContext<ContentItem>)context);
+
+    public void Describe(DescribeContext<ContentItem> context)
+    {
+        context.For<LocalizedContentItemIndex>()
+            .When(contentItem => contentItem.Has<LocalizationPart>() || _partRemoved.Contains(contentItem.ContentItemId))
+            .Map(contentItem =>
+            {
+                // If the content item was removed, a record is still added.
+                if (!contentItem.Published && !contentItem.Latest && !_itemRemoved.Contains(contentItem))
+                {
+                    return null;
+                }
+
+                // If the part was removed from the type definition, a record is still added.
+                var partRemoved = _partRemoved.Contains(contentItem.ContentItemId);
+
+                var part = contentItem.As<LocalizationPart>();
+                if (!partRemoved && (part == null || string.IsNullOrEmpty(part.LocalizationSet) || part.Culture == null))
+                {
+                    return null;
+                }
+
+                // If the part was removed, a record is still added but with a null Culture.
+                return new LocalizedContentItemIndex
+                {
+                    Culture = !partRemoved ? part.Culture.ToLowerInvariant() : null,
+                    LocalizationSet = !partRemoved ? part.LocalizationSet : null,
+                    ContentItemId = contentItem.ContentItemId,
+                    Published = contentItem.Published,
+                    Latest = contentItem.Latest
+                };
+            });
+    }
+}
